renamed files to be lower letters
