+++ main.py
@@ -5,7 +5,7 @@ from utils.data_loader import load_data
 from utils.config_loader import load_config
 from utils.logger import Logger, FileLogger
 from utils.model_manager import run_models_and_get_best, get_model_instance, train_and_predict_with_best
-from Model.Pipeline.pipeline import DataPipeline
+from model.pipeline.pipeline import DataPipeline
 
 logger = Logger()
 
@@ -27,7 +27,7 @@ def main():
     logger.log("Data loaded successfully.\n")
 
     pipeline = DataPipeline(config["bots_and_scores"], turns_df, games_df)
-    logger.log("Pipeline initialized.")
+    logger.log("pipeline initialized.")
     processed_train_df = pipeline.process_train_data(train_df)
     processed_test_df = pipeline.process_test_data(test_df)
     logger.log("Data processing completed.\n")
@@ -37,12 +37,12 @@ def main():
         model_name: get_model_instance(model_details["class"])
         for model_name, model_details in config["models"].items()
     }
-    logger.log("Models initialized.")
+    logger.log("models initialized.")
 
     # Run models and find the best model
     results_df, best_model = run_models_and_get_best(processed_train_df, models, config, logger)
-    logger.log("Model evaluation completed.\n")
-    logger.log("Model Performance Summary:")
+    logger.log("model evaluation completed.\n")
+    logger.log("model Performance Summary:")
     logger.log(results_df.to_string(index=False) + "\n")
 
     # Train and predict with the best model
