+++ model/pipeline/pipeline.py
@@ -1,15 +1,16 @@
 import pandas as pd
 from sklearn.pipeline import Pipeline
-from Model.Pipeline.Transformers.Data_Cleaning.reorder_columns import ReorderColumns
-from Model.Pipeline.Transformers.Feature_Engineering.add_is_first_play import AddIsFirstPlay
-from Model.Pipeline.Transformers.Feature_Engineering.add_bot_difficulty import AddBotDifficulty
-from Model.Pipeline.Transformers.Feature_Engineering.add_mean_rack_usage import AddMeanRackUsage
-from Model.Pipeline.Transformers.Data_Cleaning.drop_features_from_train import DropFeaturesFromTrain
-from Model.Pipeline.Transformers.Data_Cleaning.split_and_trasform_train import SplitAndTransformTrain
-from Model.Pipeline.Transformers.Data_Cleaning.merge_features_from_games import MergeFeaturesFromGames
-from Model.Pipeline.Transformers.Feature_Engineering.add_negative_end_reason import AddNegativeEndReason
-from Model.Pipeline.Transformers.Data_Cleaning.encode_categorial_features import EncodeCategoricalFeatures
-from Model.Pipeline.Transformers.Feature_Engineering.add_mean_average_letter_score import AddMeanAverageLetterScore
+
+from model.pipeline.transformers.data_cleaning.reorder_columns import ReorderColumns
+from model.pipeline.transformers.feature_engineering.add_is_first_play import AddIsFirstPlay
+from model.pipeline.transformers.feature_engineering.add_bot_difficulty import AddBotDifficulty
+from model.pipeline.transformers.feature_engineering.add_mean_rack_usage import AddMeanRackUsage
+from model.pipeline.transformers.data_cleaning.drop_features_from_train import DropFeaturesFromTrain
+from model.pipeline.transformers.data_cleaning.split_and_trasform_train import SplitAndTransformTrain
+from model.pipeline.transformers.data_cleaning.merge_features_from_games import MergeFeaturesFromGames
+from model.pipeline.transformers.feature_engineering.add_negative_end_reason import AddNegativeEndReason
+from model.pipeline.transformers.data_cleaning.encode_categorial_features import EncodeCategoricalFeatures
+from model.pipeline.transformers.feature_engineering.add_mean_average_letter_score import AddMeanAverageLetterScore
 
 
 class DataPipeline:
