+++ model/pipeline/transformers/abstract_tranformer.py
