+++ model/pipeline/transformers/feature_engineering/add_bot_difficulty.py
@@ -1,4 +1,4 @@
-from Model.Pipeline.Transformers.abstract_tranformer import AbstractTransformer
+from model.pipeline.transformers.abstract_tranformer import AbstractTransformer
 from typing import Dict
