+++ model/pipeline/transformers/feature_engineering/add_mean_rack_usage.py
@@ -1,4 +1,4 @@
-from Model.Pipeline.Transformers.abstract_tranformer import AbstractTransformer
+from model.pipeline.transformers.abstract_tranformer import AbstractTransformer
 from typing import Dict
 import pandas as pd
